problem: can't save role
milestone: query form is refactored as a form of fields.

NOTE / TAG: This version refactored the query-form, which shows pretty clear
that for semantics tier, there is nothing but rows & record to be
presented. (query form are form of record, with fields specifying
conditions)

Typescript checking is essential to refactor